Use plain spaces instead of tabs for indentation
If the code contains tab characters, it is harder for people to directly
copy and paste into shell to try some stuff...  And vim has a 'smarttab'
option that currently only works well with 'expandtab' option being set.
If the maintainers add new options such that *align with space, indent
with tab* is possible, I will be more willing to use tab characters to
indent.